+++ remote.rb
@@ -2,6 +2,10 @@ module Remote
   class Br
     class << self
 
+      def update
+        `git remote update`
+      end
+
       def exists? name
         raw_list.detect{|br_name| br_name.include? name } != nil
       end